issue #133 - revert Budget Balancing parts of PR #127
