+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 Unreleased Changes
 ------------------
 
+**This release includes database migrations.**
+
 * `Issue #118 <https://github.com/jantman/biweeklybudget/issues/118>`_ - PR to fix bugs in the
   `wishlist <https://github.com/Jaymon/wishlist>`_ dependency package, and vendor that patched
   version in under ``biweeklybudget.vendored.wishlist``.
@@ -11,10 +13,10 @@ Unreleased Changes
   git requirements (ofxclient and ofxparse) that seem unmaintained or inactive, so we can install via ``pip``.
 * `Issue #115 <https://github.com/jantman/biweeklybudget/issues/115>`_ - In Transactions view, add ability to filter by budget.
 * Change ``BiweeklyPayPeriod`` class to never convert to floats (always use decimal.Decimal types).
-* **Experimental** - `Issue #117 <https://github.com/jantman/biweeklybudget/issues/117>`_ - Experimental, initial support for pushbutton balancing of budgets in the most recent previous payperiod. This feature will transfer funds between periodic budgets in the pay period attempting to get them all to a zero remaining amount. Once as many periodic budgets are zeroed as possible, a user-selected standing budget will be used to transfer excess funds to or cover negative budgets from. Finally, funds will be transferred to or from the standing budget to result in the overall remaining amount for the pay period being zero. This feature is new and experimental, and the confirmation dialog should be closely checked for accuracy and correctness.
 * `Issue #124 <https://github.com/jantman/biweeklybudget/issues/124>`_ - Major changes to the ``ofxgetter`` and ``ofxbackfiller`` console scripts; centralize all database access in them to the new ``biweeklybudget.ofxapi.local.OfxApiLocal`` class and allow these scripts to function remotely, interacting with the ReST API instead of requiring direct database access.
 * `Issue #123 <https://github.com/jantman/biweeklybudget/issues/123>`_ - Modify the Credit Payoffs view to allow removal of Increase and Onetime Payment settings lines.
 * `Issue #131 <https://github.com/jantman/biweeklybudget/issues/131>`_ - Add better example data for screenshots.
+* `Issue #117 <https://github.com/jantman/biweeklybudget/issues/117>`_ and `#133 <https://github.com/jantman/biweeklybudget/issues/133>`_ - Implement and then revert out a failed attempt at automatic balancing of budgets in the previous pay period.
 * `Issue #114 <https://github.com/jantman/biweeklybudget/issues/114>`_
 
   * Add ``transfer_id`` field and ``transfer`` relationship to Transaction model, to link the halves of budget transfer transactions in the database. The alembic migration for this release iterates all Transactions in the database, and populates these links based on inferences of the description, date, account_id and notes fields of sequential pairs of Transactions. (Note: this migration would likely miss some links if two transfers were created simultaneously, and ended up with the Transaction IDs interleaved).
