+++ biweeklybudget/flaskapp/static/js/budgets_modal.js
@@ -68,7 +68,6 @@ function budgetModalDivForm() {
         .addCurrency('budget_frm_current_balance', 'current_balance', 'Current Balance', { groupHtml: 'style="display: none;"' })
         .addCheckbox('budget_frm_active', 'is_active', 'Active?', true)
         .addCheckbox('budget_frm_income', 'is_income', 'Income?')
-        .addCheckbox('budget_frm_skip_balance', 'skip_balance', 'Skip Balancing?')
         .render();
 }
 
@@ -100,11 +99,6 @@ function budgetModalDivFillAndShow(msg) {
         $('#budget_frm_income').prop('checked', true);
     } else {
         $('#budget_frm_income').prop('checked', false);
-    }
-    if(msg['skip_balance'] === true) {
-        $('#budget_frm_skip_balance').prop('checked', true);
-    } else {
-        $('#budget_frm_skip_balance').prop('checked', false);
     }
     $("#modalDiv").modal('show');
 }
