--- docs/source/biweeklybudget.budget_balancer.rst
@@ -1,7 +1,0 @@
-biweeklybudget\.budget\_balancer module
-=======================================
-
-.. automodule:: biweeklybudget.budget_balancer
-    :members:
-    :undoc-members:
-    :show-inheritance:
