+++ docs/source/biweeklybudget.models.utils.rst
@@ -1,0 +1,7 @@
+biweeklybudget\.models\.utils module
+====================================
+
+.. automodule:: biweeklybudget.models.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
